+++ src/srcp-descr.c
@@ -1,7 +1,7 @@
 /* $Id$ */
 
-/* 
- * Vorliegende Software unterliegt der General Public License, 
+/*
+ * Vorliegende Software unterliegt der General Public License,
  * Version 2, 1991. (c) Matthias Trute, 2000-2001.
  *
  */
@@ -15,8 +15,8 @@ int startup_DESCRIPTION(void)
   return 0;
 }
 
-int describeBus(int bus, char *reply) {
-      sprintf(reply, "%lu.%.3lu 100 INFO %d DESCRIPTION %s\n",
+int describeBus(long int bus, char *reply) {
+      sprintf(reply, "%lu.%.3lu 100 INFO %ld DESCRIPTION %s\n",
         busses[bus].power_change_time.tv_sec,  busses[bus].power_change_time.tv_usec/1000,
         bus, busses[bus].description);
       return SRCP_INFO;
