+++ src/srcp-fb.h
@@ -1,7 +1,7 @@
 /* $Id$ */
 
-/* 
- * Vorliegende Software unterliegt der General Public License, 
+/*
+ * Vorliegende Software unterliegt der General Public License,
  * Version 2, 1991. (c) Matthias Trute, 2000-2001.
  *
  */
@@ -32,15 +32,15 @@ typedef struct _RESET_FB
 } reset_FB;
 
 int startup_FB(void);
-int init_FB(int bus, int number);
-int get_number_fb(int bus);
-int initFB(int busnumber, int addr, const char protocol, int index);
+int init_FB(long int bus, int number);
+int get_number_fb(long int bus);
+int initFB(long int busnumber, int addr, const char protocol, int index);
 
-int getFB(int bus, int port, struct timeval *time, int *value);
-void updateFB(int bus, int port, int value);
-int setFBmodul(int bus, int mod, int values);
-int infoFB(int bus, int port, char *msg);
-int describeFB(int bus, int addr, char *reply);
-void check_reset_fb(int busnumber);
-void set_min_time(int busnumber, int mt);
+int getFB(long int bus, int port, struct timeval *time, int *value);
+void updateFB(long int bus, int port, int value);
+int setFBmodul(long int bus, int mod, int values);
+int infoFB(long int bus, int port, char *msg);
+int describeFB(long int bus, int addr, char *reply);
+void check_reset_fb(long int busnumber);
+void set_min_time(long int busnumber, int mt);
 #endif
