+++ src/zimo.h
@@ -9,10 +9,10 @@ typedef struct _zimo_DATA {
     int number_fb;
 } zimo_DATA;
 
-int readconfig_zimo(xmlDocPtr doc, xmlNodePtr node, int busnumber);
+int readconfig_zimo(xmlDocPtr doc, xmlNodePtr node, long int busnumber);
 int init_linezimo(char *);
-int init_bus_zimo(int );
-int term_bus_zimo(int );
+long int init_bus_zimo(long int);
+long int term_bus_zimo(long int);
 int getDescription_zimo(char *reply);
 void* thr_sendrec_zimo(void *);
